Add 'ui.menu' subpackage
