+++ src/edu/colorado/csdms/wmt/client/ui/Perspective.java
@@ -44,10 +44,10 @@ import edu.colorado.csdms.wmt.client.control.DataManager;
 import edu.colorado.csdms.wmt.client.data.LabelJSO;
 import edu.colorado.csdms.wmt.client.ui.dialog.ComponentInfoDialogBox;
 import edu.colorado.csdms.wmt.client.ui.dialog.OpenDialogBox;
+import edu.colorado.csdms.wmt.client.ui.menu.ComponentSelectionMenu;
+import edu.colorado.csdms.wmt.client.ui.menu.LabelsMenu;
 import edu.colorado.csdms.wmt.client.ui.panel.ModelActionPanel;
 import edu.colorado.csdms.wmt.client.ui.panel.UserPanel;
-import edu.colorado.csdms.wmt.client.ui.widgets.ComponentSelectionMenu;
-import edu.colorado.csdms.wmt.client.ui.widgets.LabelsMenu;
 
 /**
  * Defines the initial layout of views (a perspective, in Eclipse parlance)
