+++ src/edu/colorado/csdms/wmt/client/ui/menu/LabelsMenu.java
@@ -21,21 +21,26 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package edu.colorado.csdms.wmt.client.ui.widgets;
+package edu.colorado.csdms.wmt.client.ui.menu;
 
 import java.util.Map;
 import java.util.Map.Entry;
 
 import com.google.gwt.event.dom.client.ClickEvent;
 import com.google.gwt.event.dom.client.ClickHandler;
+import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.ui.CheckBox;
+import com.google.gwt.user.client.ui.HTML;
 import com.google.gwt.user.client.ui.PopupPanel;
 import com.google.gwt.user.client.ui.ScrollPanel;
 import com.google.gwt.user.client.ui.VerticalPanel;
 
 import edu.colorado.csdms.wmt.client.Constants;
 import edu.colorado.csdms.wmt.client.control.DataManager;
+import edu.colorado.csdms.wmt.client.control.DataTransfer;
 import edu.colorado.csdms.wmt.client.data.LabelJSO;
+import edu.colorado.csdms.wmt.client.ui.dialog.LabelDialogBox;
+import edu.colorado.csdms.wmt.client.ui.handler.DialogCancelHandler;
 
 /**
  * Encapsulates an alphabetized, scrollable list of labels used to tag and
@@ -43,21 +48,24 @@ import edu.colorado.csdms.wmt.client.data.LabelJSO;
  * 
  * @author Mark Piper (mark.piper@colorado.edu)
  */
-public class LabelsSaveModelMenu extends PopupPanel {
+public class LabelsMenu extends PopupPanel {
 
   private DataManager data;
   private VerticalPanel labelPanel;
-
+  private HTML addNewHtml;
+  private HTML deleteHtml;
+  
   /**
-   * Makes a new {@link LabelsSaveModelMenu}.
+   * Makes a new {@link LabelsMenu}.
    * 
    * @param data the DataManager object for the WMT session
    */
-  public LabelsSaveModelMenu(DataManager data) {
+  public LabelsMenu(DataManager data) {
     
     super(true); // autohide
     this.data = data;
     this.setStyleName("wmt-PopupPanel");
+    data.getPerspective().setLabelsMenu(this);
 
     // A VerticalPanel for the menu items. (PopupPanels have only one child.)
     VerticalPanel menu = new VerticalPanel();
@@ -71,10 +79,25 @@ public class LabelsSaveModelMenu extends PopupPanel {
 
     // Populate the menu with the stored model labels and their values.
     populateMenu();
+
+    // These items are always visible on the bottom of the menu.
+    HTML separator = new HTML("");
+    separator.setStyleName("wmt-PopupPanelSeparator");
+    addNewHtml = new HTML("Add new label");
+    addNewHtml.setStyleName("wmt-PopupPanelItem");
+    deleteHtml = new HTML("Delete label");
+    deleteHtml.setStyleName("wmt-PopupPanelItem");
+    menu.add(separator);
+    menu.add(addNewHtml);
+    menu.add(deleteHtml);
+    
+    // Show a SuggestBox to add or delete a label.
+    addNewHtml.addClickHandler(new LabelAddRemoveHandler(data, "Add"));
+    deleteHtml.addClickHandler(new LabelAddRemoveHandler(data, "Delete"));    
   }
   
   /**
-   * A helper that loads the {@link LabelsSaveModelMenu} with {@link CheckBox} labels.
+   * A helper that loads the {@link LabelsMenu} with {@link CheckBox} labels.
    * Each CheckBox has a handler that maps the selection state of the box to the
    * labels variable stored in the {@link DataManager}.
    */
@@ -102,6 +125,17 @@ public class LabelsSaveModelMenu extends PopupPanel {
       }
     }
   }
+  
+  /**
+   * Deselects all labels except for the username label. This is the default
+   * state for the {@link LabelsMenu}.
+   */
+  public void resetSelections() {
+    for (Map.Entry<String, LabelJSO> entry : data.modelLabels.entrySet()) {
+      entry.getValue().isSelected(
+          entry.getKey().equals(data.security.getWmtUsername()));
+    }
+  }
 
   /**
    * Handles actions when the selection state of a label changes.
@@ -121,7 +155,69 @@ public class LabelsSaveModelMenu extends PopupPanel {
       CheckBox labelBox = (CheckBox) event.getSource();
       entry.getValue().isSelected(labelBox.getValue());
       data.updateModelSaveState(false);
-    } 
+    }
   }
 
+  /**
+   * Handles adding and deleting model labels using a {@link LabelDialogBox}.
+   */
+  public class LabelAddRemoveHandler implements ClickHandler {
+
+    private DataManager data;
+    private String type;
+
+    /**
+     * Creates a new {@link LabelAddRemoveHandler}.
+     * 
+     * @param data the DataManager object for the WMT session
+     * @param type the event type, currently "Add" or "Delete"
+     */
+    public LabelAddRemoveHandler(DataManager data, String type) {
+      this.data = data;
+      this.type = type; // use sentence case
+    }
+    
+    @Override
+    public void onClick(ClickEvent event) {
+      
+      final LabelDialogBox box = new LabelDialogBox(data);
+      
+      box.getChoicePanel().getCancelButton().addClickHandler(
+          new DialogCancelHandler(box));
+      box.getChoicePanel().getOkButton().setHTML(Constants.FA_TAGS + type);
+      box.getChoicePanel().getOkButton().addClickHandler(new ClickHandler() {
+        @Override
+        public void onClick(ClickEvent event) {
+          String label = box.getSuggestBox().getText();
+          if (type.equalsIgnoreCase("add")) {
+            if (data.modelLabels.containsKey(label)) {
+              Window.alert(Constants.ADD_LABEL_ERR);
+            } else {
+              DataTransfer.addLabel(data, label);
+            }
+          } else if (type.equalsIgnoreCase("delete")) {
+            LabelJSO jso = data.modelLabels.get(label);
+            if (jso != null) {
+              if (!data.security.getWmtUsername().matches(jso.getOwner())) {
+                Window.alert(Constants.DELETE_LABEL_ERR);
+              } else {
+                if (jso.isSelected()) {
+                  data.updateModelSaveState(false);
+                }
+                DataTransfer.deleteLabel(data, jso.getId());
+              }
+            }
+          }
+          box.hide();
+        }
+      });
+
+      if (type.equalsIgnoreCase("add")) {
+        box.showRelativeTo(addNewHtml);
+      } else if (type.equalsIgnoreCase("delete")) {
+        box.showRelativeTo(deleteHtml);
+      }
+      box.getSuggestBox().setFocus(true);
+    }
+  }
 }
