+++ src/edu/colorado/csdms/wmt/client/ui/menu/ViewInputFilesMenu.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package edu.colorado.csdms.wmt.client.ui.widgets;
+package edu.colorado.csdms.wmt.client.ui.menu;
 
 import java.util.Iterator;
