+++ src/edu/colorado/csdms/wmt/client/ui/menu/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Menus used in the WMT client interface.
+ * 
+ * @author Mark Piper (mark.piper@colorado.edu)
+ */
+package edu.colorado.csdms.wmt.client.ui.menu;
