+++ src/edu/colorado/csdms/wmt/client/ui/panel/ParameterActionPanel.java
@@ -36,7 +36,7 @@ import edu.colorado.csdms.wmt.client.Constants;
 import edu.colorado.csdms.wmt.client.control.DataManager;
 import edu.colorado.csdms.wmt.client.ui.dialog.ComponentInfoDialogBox;
 import edu.colorado.csdms.wmt.client.ui.handler.ParameterActionPanelResetHandler;
-import edu.colorado.csdms.wmt.client.ui.widgets.ViewInputFilesMenu;
+import edu.colorado.csdms.wmt.client.ui.menu.ViewInputFilesMenu;
 
 /**
  * Makes a row of action buttons ("Reset", "View input files", etc.) for working
